changed Point to double

+++ include/Point.hpp
@@ -9,13 +9,13 @@ class Plane;
 
 class Point{
 private:
-    int x, y, z;
+    double x, y, z;
 public:
     Point();
-    Point(int x_coordinate, int y_coordinate, int z_coordinate);
-    int X() const;
-    int Y() const;
-    int Z() const;
+    Point(double x_coordinate, double y_coordinate, double z_coordinate);
+    double X() const;
+    double Y() const;
+    double Z() const;
     bool on_outer_side(const Point A, const Point B, const Point C) const;   //A, B, C represent a triangle, whose outer side is the one where the points are seen in this order clockwise
     bool on_inner_side(Point A, Point B, Point C) const;
     bool under(Plane &plane) const;
