+++ src/Point.cpp
@@ -5,10 +5,10 @@
 #include"Plane.hpp"
 
 Point::Point() : x{0}, y{0}, z{0} {}
-Point::Point(int x_coordinate, int y_coordinate, int z_coordinate) : x{x_coordinate}, y{y_coordinate}, z{z_coordinate} {}
-int Point::X() const {return x;}
-int Point::Y() const {return y;}
-int Point::Z() const {return z;}
+Point::Point(double x_coordinate, double y_coordinate, double z_coordinate) : x{x_coordinate}, y{y_coordinate}, z{z_coordinate} {}
+double Point::X() const {return x;}
+double Point::Y() const {return y;}
+double Point::Z() const {return z;}
 
 bool Point::on_outer_side(const Point A, const Point B, const Point C) const {
     Plane plane(A, B, C);
@@ -68,11 +68,11 @@ bool Point::inside(Point *A, Point *B, Point *C, Plane &plane) const {
     Vector v1 = *B - *A;
     Vector v2 = *this - *A;
 
-    int dot00 = v0.dot(v0);
-    int dot01 = v0.dot(v1);
-    int dot02 = v0.dot(v2);
-    int dot11 = v1.dot(v1);
-    int dot12 = v1.dot(v2);
+    double dot00 = v0.dot(v0);
+    double dot01 = v0.dot(v1);
+    double dot02 = v0.dot(v2);
+    double dot11 = v1.dot(v1);
+    double dot12 = v1.dot(v2);
 
     double denom = (dot00 * dot11 - dot01 * dot01);
     double alpha = (dot11 * dot02 - dot01 * dot12) / denom;
@@ -82,9 +82,9 @@ bool Point::inside(Point *A, Point *B, Point *C, Plane &plane) const {
     return alpha >= 0.0f && betha >= 0.0f && gamma >= 0.0f;
 }
 Vector Point::operator-(const Point &that) const {
-    int u = this->x - that.X();
-    int v = this->y - that.Y();
-    int w = this->z - that.Z();
+    double u = this->x - that.X();
+    double v = this->y - that.Y();
+    double w = this->z - that.Z();
     return Vector(u, v, w);
 }
 std::ostream &operator<<(std::ostream &os, const Point &P) {
@@ -92,8 +92,8 @@ std::ostream &operator<<(std::ostream &os, const Point &P) {
 }
 
 Point Point::operator+(const Vector &that) const{
-    int x = this->x + that.X();
-    int y = this->y + that.Y();
-    int z = this->z + that.Z();
+    double x = this->x + that.X();
+    double y = this->y + that.Y();
+    double z = this->z + that.Z();
     return Point(x, y, z);
 }
